ripplit/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
xrpl-py>=2.0
pydantic>=2.0
httpx>=0.27
